add development & CI test profiles
Added two test profiles: one for development (devel) and one for
running the integration tests as part of the CI.

Currently the only difference between the two is the result of the
coverage test. Specifically, in the `ci` profile the test fails
if current_coverage > previous_coverage, while in the `devel`
profile the coverage file is updated with the new value.

Signed-off-by: Andreea Florescu <fandree@amazon.com>