vcpu: Return proper vector when receiving EOI
The vector value returned along with End Of Interrupt (EOI) VM exit
is important as it indicates a userspace IOAPIC which interrupt needs
to be deasserted.

About KVM_EXIT_IOAPIC_EOI, KVM documentation mentions:
```
/* KVM_EXIT_IOAPIC_EOI */
struct {
    __u8 vector;
} eoi;

Indicates that the VCPU's in-kernel local APIC received an EOI for a
level-triggered IOAPIC interrupt. This exit only triggers when the
IOAPIC is implemented in userspace (i.e. KVM_CAP_SPLIT_IRQCHIP is
enabled); the userspace IOAPIC should process the EOI and retrigger
the interrupt if it is still asserted. Vector is the LAPIC interrupt
vector for which the EOI was received.
```

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>