use rust-vmm-ci for running the integration tests
Use rust-vmm-ci as a git submodule. Currently this points to
my fork of rust-vmm-ci. Will update once the rust-vmm-ci is merged.

Added coverage_config.json file that is used by test_coverage.py
to configure parameters of the coverage test like current coverage,
additional files to exclude when computing the coverage and extra
crates features to run the tests for.

Signed-off-by: Andreea Florescu <fandree@amazon.com>