Add support for KVM_SET_GSI_ROUTING ioctl
By adding the support for the ioctl KVM_SET_GSI_ROUTING, this commit
will allow consumers of the kvm-ioctls crate to set custom mappings for
GSI entries.

This would let a caller of this ioctl defines that GSI number X can
trigger an interrupt by delivering an MSI directly to the local APIC or
instead deliver a pin based interrupt to the in-kernel IOAPIC.

One example that comes to mind is to use this ioctl to let a VMM
interact with the VFIO API, as it requires an eventfd to be associated
with each interrupt. By combining this ioctl with KVM_IRQFD, one could
directly map an interrupt to an eventfd managed by KVM, and associated
to a specific GSI.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>