make test_irq_chip x86 specific
The test was previously marked as x86_64 and aarch64 compatible.

On aarch64 we were testing only the error case. On older kernel
versions, creating an irqchip before creating the vCPU would return
an error. This restrictions is not in place anymore on Linux Kernel
4.15.

Since we cannot test this error path anymore on aarch64, the test is
now labeled as x86 specific.

Also updated the test to set the irq_base of the PIC and check that
this value is appropriately updated.

Signed-off-by: Andreea Florescu <fandree@amazon.com>