Add unit test for KVM_CREATE_IRQCHIP on ARM.
On ARM, KVM_CREATE_IRQCHIP is a bit more complex than on X86.
KVM_CREATE_IRQCHIP only creates VGICv2. It usually succeed with a
GICv2 hardware. But if the hardware is GICv3, it's probable that
GICv2 can't be emulated, so KVM_CREATE_IRQCHIP would fail.

An issue was created for the problem:
https://github.com/rust-vmm/kvm-ioctls/issues/79

To avoid hardware dependency, in the test we'd better check the
ability of creating VGICv2 in advance, and assert the result of
KVM_CREATE_IRQCHIP correspondingly.

Change-Id: Id17339dde28dbb4abd68bf09fe7f8f0d88c14edb
Signed-off-by: Michael Zhao <michael.zhao@arm.com>