Add support for `KVM_GET_DEVICE_ATTR` ioctl
The `KVM_GET_DEVICE_ATTR` ioctl is useful when we need to extract
the state/information of devices in the VM. In AArch64 VMs, using
this ioctl is the only method to get the vGIC states.

This commit implements the `KVM_GET_DEVICE_ATTR` ioctl with its
unit test on AArch64.

Fixes: https://github.com/rust-vmm/kvm-ioctls/issues/99

Signed-off-by: Henry Wang <Henry.Wang@arm.com>