Enable `KVM_MEM_ENCRYPT_OP` ioctl for AMD SEV
The `KVM_MEM_ENCRYPT_OP` ioctl is used to access AMD SEV (Secure
Encrypted Virtualization) feature. When enabled, memory contents of a VM
will be transparently encrypted with a key unique to that VM.

This patch added implementation, documentation and related tests for the
following SEV commands:
- KVM_SEV_INIT
- KVM_SEV_LAUNCH_START
- KVM_SEV_LAUNCH_UPDATE_DATA
- KVM_SEV_LAUNCH_MEASURE
- KVM_SEV_LAUNCH_FINISH
- KVM_SEV_LAUNCH_SECRET

Note: this work is based on enarx/ketuvim. See [1]_ for details.

[1] https://github.com/rust-vmm/community/issues/62

Signed-off-by: Kailun Qin <kailun.qin@hotmail.com>