vcpu: Test Case for num_entries too small CpuId
Kernel returns an error `E2BIG` if the number of entries passed to
`get_cpuid2` is too small. Added a test case for that assertion. Also,
added assertion to check `errno` to `ENOMEM`, when passed number of
entries is greater than `KVM_MAX_CPUID_ENTRIES` (previously the
assertion was there only for `is_err`).

Signed-off-by: Abhijit Gadgil <gabhijit@iitbombay.org>