fix get_cpuid test
With new kernel versions the CPUID entries are no longer returned in
order. To be able to compare cpuid entries we need to first sort them.
This requires changes in kvm-bindings as well as we need first to have
the kvm_cpuid_entry2 sortable (i.e. implement Ord and PartialOrd).

Signed-off-by: Andreea Florescu <fandree@amazon.com>