Add support for coalesced MMIO (KVM_CAP_COALESCED_MMIO)
Add support for coalesced MMIO. This performance feature allows guest
writes to port and memory space to not trigger VM exits. Instead, the
kernel will write an entry into a shared ring buffer for each access,
which userspace must consume. The ring buffer is mapped at a certain
offset in the vcpu's file descriptor.

In order to enable this capability, introduce the KvmCoalescedIoRing
struct, which will act as a safe wrapper around the raw mapping of
the ring buffer. Since users may not use coalesced MMIO, or it might
not be available, store it as an Option in the VcpuFd struct.

Signed-off-by: Carlos López <carlos.lopez@suse.com>