Add `serde` support for various bindings
Add implementations of `serde::{Serialize, Deserialize}` for
kvm-bindings structures commonly used for live migration/snapshotting.
The selections of structures to add the derives to was determined by
finding the minimal set that allows both Firecracker and CloudHypervisor
to compile against this branch.

Serialization happens as opaque binary blobs via the `zerocopy` crate.
This is to support live migration across different versions of
kvm-bindings (which can potentially be using bindings generated from
different versions of the kernel headers). Since the kernel's ABI is
backward-compatible, the only changes that can happen are:

- "reserved" areas of structures acquire meaning. This is a change
  purely at the representational level, and has no impact on the
  underlying binary representation. Thus, our serialization is
  roboust to these changes (whereas simply deriving the serde traits
  would cause incompatibilities here).

- Adding new fields to the end of a structure. When deserializating, we
  compare the length of the deserialized data with the size of the
  struct, and pad with zeros/truncate as appropriate.

This allows seamless live update.

We pick the `zerocopy` crate for binary serialization as it is already
in use in the rust-vmm ecosystem (in `acpi_tables`), and it has a very
strong security stance (for example, using formal verification in their
CI). The alternative would have been using `ByteValued` from
`vm-memory`, but I do not want `kvm-bindings` to gain a dependency on
`vm-memory` (especially since `vm-memory` brings in a lot more
functionality that `kvm-bindings` does not need).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>