fix: derive `Default` for `kvm_xsave2`
`FamStructWrapper<T>` imposes a bound of `Default + FamStruct` on `T`.
However, it seems that when defining type aliases, Rust will not ensure
this bound is met until the type alias is actually used (as in a
function definition). This made the `Xsave` type defined in
`x86_64/fam.rs` essentially "unmentionable", as any attempts as using it
will result in a compile error.

Fix this by deriving `Default` for it, and also derive `Debug` for
consistency with other types.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>