Fix IO safety runtime errors in tests
If debug_assertions is enabled, Rust 1.80.0 and later will abort the
process when attempting to close a file descriptor that is already
closed[1].  This means that, in tests, where an invalid file
descriptor object is deliberately constructed, we have to avoid that
object being dropped.  Using into_raw_fd() allows the object's memory
to be freed (unlike std::mem::forget()), without attempting to close
the file descriptor.

[1]: https://github.com/rust-lang/rust/commit/cb4940645775f60d74aee2e018d6c516c5aa9ed7

Signed-off-by: Alyssa Ross <hi@alyssa.is>