fix: use correct ioctl wrapper for `create_device`
Use `ioctl_with_mut_ref` instead of `ioctl_with_ref`
in the `create_device` method as it needs to write to the
`kvm_create_device` struct passed to it. This incorrect
usage of `ioctl_with_ref` causes newer versions of Rust compiler
(1.82 and above) to treat the `kvm_create_device` struct as read-only
and bypass following reads from it. This optimization lead to incorrect
value being passed to the `File::from_raw_fd` call.

Signed-off-by: Egor Lazarchuk <yegorlz@amazon.co.uk>